+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [tailwindcss configuration css class utilities](https://tailwindcss.com/docs/guides/nextjs)
 - test tailwind configuration on index.ts referencing [with tailwind sample](https://github.com/vercel/next.js/blob/canary/examples/with-tailwindcss/pages/index.js)
 - [add ant design for built UI components](https://ant.design/components/overview/)
+- add [react hook form](https://react-hook-form.com) for forms implementation